The Temple of Elemental Expeditions
Expeditions (modules if you will) that the party can go to, meeting the
challenges to earn XP.

This is just T1, really. Ts 2-4 will start dealing with the results of
taking damage, flavour text of the challenges, themes for the
expeditions (to lean towards magic for a wizard lair, or dos for a
Giant stedding, for example), and gold pieces.